Shiny interface to the pairsD3 function

Opens a shiny GUI to facilitate interaction with the pairsD3 function

shinypairs(x, group = NULL, subset = NULL, labels = NULL)

Arguments

x
the coordinates of points given as numeric columns of a matrix or data frame. Logical and factor columns are converted to numeric in the same way that data.matrix does.

group
a optional vector specifying the group each observation belongs to. Used for tooltips and colouring the observations.

subset
an optional vector specifying a subset of observations to be used for plotting. Useful when you have a large number of observations, you can specify a random subset.

labels
the names of the variables (column names of x used by default).

Examples

data(iris)
# NOT RUN {
shinypairs(iris)
# }