D3 scatterplot matrices

An interactive matrix of scatterplots is produced.

pairsD3(x, group = NULL, subset = NULL, labels = NULL, cex = 3,
  width = NULL, col = NULL, big = FALSE, theme = "colour",
  opacity = 0.9, tooltip = NULL, leftmar = 35, topmar = 2)

Arguments

x
the coordinates of points given as numeric columns of a matrix or data frame. Logical and factor columns are converted to numeric in the same way that data.matrix does.

group
a optional vector specifying the group each observation belongs to. Used for tooltips and colouring the observations.

subset
an optional vector specifying a subset of observations to be used for plotting. Useful when you have a large number of observations, you can specify a random subset.

labels
the names of the variables (column names of x used by default).

cex
the magnification of the plotting symbol (default=3)

width
the width (and height) of the plot when viewed externally.

col
an optional (hex) colour for each of the levels in the group vector.

big
a logical parameter. Prevents inadvertent plotting of huge data sets. Default limit is 10 variables, to plot more than 10 set big=TRUE.

theme
a character parameter specifying whether the theme should be colour colour (default) or black and white bw.

opacity
numeric between 0 and 1. The opacity of the plotting symbols (default 0.9).

tooltip
an optional vector with the tool tip to be displayed when hovering over an observation. You can include basic html.

leftmar
space on the left margin

topmar
space on the bottom margin

Examples

data(iris)
# NOT RUN {
pairsD3(iris[,1:4],group=iris[,5],
         labels=gsub(pattern = "\\.",replacement = " ", names(iris)))
# }